+++ fec/data/templates/layouts/widgets.jinja
@@ -25,9 +25,6 @@
     {# WEBMANAGER_EMAIL = '{{ WEBMANAGER_EMAIL }}'; #}
     {# CANONICAL_BASE = '{{ CANONICAL_BASE }}'; #}
     {# TRANSITION_URL = '{{ TRANSITION_URL }}'; #}
-    var ELECTION_YEARS = ['2024','2022','2020','2018','2016','2014','2012','2010','2008','2006','2004','2002','2000','1998','1996','1994','1992','1990','1988','1986','1984','1982','1980'];
-    var ELECTION_YEAR = 2012;
-    var TEST = 'YAY!';
 
     function trackMetric(name, imagesSelector) {
       if (imagesSelector) {
