+++ fec/fec/static/js/widgets/_aggregate-totals-v0.js
